requirements.txt
playwright
pytest
pytest-html
pytest-xdist
pytest-rerunfailures
python-dotenv
